Make the network rest after test.
There where differences in the counters between the custom and snntorch
version.

Starting the test from a different image in the custom version led to different
classification results. This means that during the temporal evolution of the
test something changes with respect to the expected behaviour.

The error was that the network was not reset after the testing phase: this
implied that at every image the network started from a different state.

Inserting the rest of the network solved the problem.

There is still a slight difference in the output counters and this affect the
accuracy.