Update the datapath.
1) Move the comparison after the voltage register: compare v and v_th instead of
v_value and v_th

2) Invert the inputs of the v_mux: it doesn't change anything in practice, but
during the modification I inverted them, so everything is switched now.

3) Modify the exc_weight bit-width alignment: the neuron exposes a signed
weight. I expect it to be treated as signed, so if the MSB is 1 the resulting
number is negative. This makes me lose a bit, but I can also have neuron
bit-width = neuron bit-width. Let's see if the modification is ok.