Single bit signal dedicated to replicate sign.
VHDL 93 or 2002, or the mixed 93-2008 used by xvhdl by default gave me error if
I tried to connect a subportion of in_weight and exc_weight in the port map. It
gave problems even with the chosen solution, because (others => ) wants a
constant argument.

I switch the VHDL compiler to the 2008 version. It doesn't give problems.