Add the possibility to specify debug signal list.
This can speed uo the debug: the first time the debug can be run simply and the
user can annotate the signals he wants to bring outside. From the second run the
signals list can be passed directly to the class to track them without the need
to manually re-insert them every time.