Correct error when generating multiplexer.
I used the same multiplexer only if n_exc_inputs ) n_inh_inputs. The problem is
that for example n_exc_inputs = 4 and n_inh_inputs = 3 require the same
multiplexer, that is a 4 to 1, even if they are different. So I need to compare
ceil_pow2(n_exc_inputs) and ceil_po2(n_inh_inputs)